extract folder page into it's own component
